backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for daydream2"
authors = ["Andrew Melbourne <melbourneandrew@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-multipart = "^0.0.20"
pydantic = "^2.11.7"
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0.42"
psycopg2-binary = "^2.9.10"
alembic = "^1.16.4"
groq = "^0.13.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

